refactor: catalogue utils to simplify testing apps with multiple schemas
Ports https://github.com/garden-co/jazz/pull/1065 into the new core
