+++ examples/docs/todo-server-ts/tests/integration.test.ts
@@ -12,6 +12,8 @@ import { mkdtempSync } from "node:fs";
 import { join } from "node:path";
 import { WebSocket as UndiciWebSocket } from "undici";
 import { deploy, startLocalJazzServer, type LocalJazzServerHandle } from "jazz-tools/testing";
+import permissions from "../permissions.js";
+import { app } from "../schema.js";
 import {
   createServer,
   startServer,
@@ -34,7 +36,8 @@ describe("Todo Server Integration", () => {
       serverUrl: upstream.url,
       appId: upstream.appId,
       adminSecret: upstream.adminSecret,
-      schemaDir: join(import.meta.dirname, ".."),
+      schema: app,
+      permissions,
     });
 
     // Create server with temp persistent storage plus an ephemeral upstream server.
@@ -206,7 +209,8 @@ describe("Todo Server Integration", () => {
         serverUrl: coldStartUpstream.url,
         appId: coldStartUpstream.appId,
         adminSecret: coldStartUpstream.adminSecret,
-        schemaDir: join(import.meta.dirname, ".."),
+        schema: app,
+        permissions,
       });
 
       // Use a shared data path so both server instances see the same Fjall file
@@ -286,7 +290,8 @@ describe("Todo Server Integration", () => {
         serverUrl: sseUpstream.url,
         appId: sseUpstream.appId,
         adminSecret: sseUpstream.adminSecret,
-        schemaDir: join(import.meta.dirname, ".."),
+        schema: app,
+        permissions,
       });
       const sseServer = await startServer(
         await createServer({
