+++ examples/todo-client-localfirst-ts/tests/browser/global-setup.ts
@@ -1,5 +1,6 @@
-import { join } from "node:path";
 import { deploy, startLocalJazzServer, type LocalJazzServerHandle } from "jazz-tools/testing";
+import permissions from "../../permissions.js";
+import { app } from "../../schema.js";
 import { TEST_PORT, JWT_SECRET, ADMIN_SECRET, APP_ID } from "./test-constants.js";
 
 export { TEST_PORT, JWT_SECRET, ADMIN_SECRET, APP_ID };
@@ -23,7 +24,8 @@ export async function setup(): Promise<void> {
     serverUrl: serverHandle.url,
     appId: serverHandle.appId,
     adminSecret: serverHandle.adminSecret,
-    schemaDir: join(import.meta.dirname, "../.."),
+    schema: app,
+    permissions,
   });
 }
