+++ packages/jazz-tools/package.json
@@ -124,6 +124,7 @@
     "build:runtime": "node scripts/clean-dist.mjs && tsc",
     "build:svelte": "svelte-package -i src/svelte -o dist/svelte --tsconfig tsconfig.svelte.json && node scripts/cleanup-generated-src-types.mjs",
     "build:solid": "vite build --config vite.config.solid.ts --mode solid-dom && vite build --config vite.config.solid.ts --mode solid-ssr && tsc --project tsconfig.solid.json --emitDeclarationOnly",
+    "check": "tsc --noEmit",
     "test": "pnpm test:node || status=1; pnpm test:react || status=1; pnpm test:svelte || status=1; pnpm test:solid || status=1; pnpm typecheck:react-native || status=1; exit ${status:-0}",
     "test:node": "vitest run --config vitest.config.ts",
     "test:react": "vitest run --config vitest.config.react.ts",
