+++ packages/jazz-tools/src/better-auth-adapter/index.test.ts
@@ -2,7 +2,8 @@ import { join } from "node:path";
 import { afterEach, beforeEach, describe, expect, it, test, vi } from "vitest";
 import { betterAuth, type BetterAuthOptions, type DBAdapter } from "better-auth";
 import { createJazzContext, type JazzContext } from "../backend/index.js";
-import { deploy, startLocalJazzServer, type LocalJazzServerHandle } from "../testing/index.js";
+import { startLocalJazzServer, type LocalJazzServerHandle } from "../testing/index.js";
+import { deploy as deployProject } from "../dev/catalogue-project.js";
 import { wasmSchema as wasmSchemaExample } from "./fixtures/schema.js";
 import { jazzAdapter } from "./index.js";
 
@@ -17,7 +18,7 @@ describe("jazzAdapter", () => {
         backendSecret: "backend-secret-for-adapter-methods",
       });
 
-      await deploy({
+      await deployProject({
         serverUrl: server.url,
         appId: server.appId,
         adminSecret: server.adminSecret,
@@ -472,7 +473,7 @@ describe("jazzAdapter", () => {
         backendSecret: "backend-secret-for-common-user-flows",
       });
 
-      await deploy({
+      await deployProject({
         serverUrl: server.url,
         appId: server.appId,
         adminSecret: server.adminSecret,
@@ -841,7 +842,7 @@ describe("jazzAdapter", () => {
         backendSecret: "backend-secret-for-better-auth-usage",
       });
 
-      await deploy({
+      await deployProject({
         serverUrl: server.url,
         appId: server.appId,
         adminSecret: server.adminSecret,
@@ -921,7 +922,7 @@ describe("jazzAdapter", () => {
         backendSecret: "backend-secret-for-integration-tests",
       });
 
-      await deploy({
+      await deployProject({
         serverUrl: server.url,
         appId: server.appId,
         adminSecret: server.adminSecret,
@@ -987,7 +988,7 @@ describe("jazzAdapter", () => {
     });
 
     test("creates and reads records through the sync server", async () => {
-      await deploy({
+      await deployProject({
         serverUrl: server.url,
         appId: server.appId,
         adminSecret: server.adminSecret,
