+++ packages/jazz-tools/src/dev/catalogue.test.ts
@@ -53,7 +53,7 @@ export default s.definePermissions(app, ({ policy, session }) => [
 }
 
 describe("dev catalogue API exports", () => {
-  it("exports project-level catalogue operations from jazz-tools/dev", async () => {
+  it("exports catalogue operations from jazz-tools/dev", async () => {
     const dev = await import("./index.js");
 
     expect(typeof dev.pushSchema).toBe("function");
@@ -142,7 +142,7 @@ describe("dev catalogue push behavior", () => {
     );
 
     const events: unknown[] = [];
-    const { deploy } = await import("./index.js");
+    const { deploy } = await import("./catalogue-project.js");
     const result = await deploy({
       appId: APP_ID,
       serverUrl: SERVER_URL,
@@ -227,7 +227,7 @@ describe("dev catalogue push behavior", () => {
     );
 
     const events: unknown[] = [];
-    const { deploy } = await import("./index.js");
+    const { deploy } = await import("./catalogue-project.js");
     const result = await deploy({
       appId: APP_ID,
       serverUrl: SERVER_URL,
@@ -320,7 +320,7 @@ describe("dev catalogue push behavior", () => {
     );
 
     const events: unknown[] = [];
-    const { deploy } = await import("./index.js");
+    const { deploy } = await import("./catalogue-project.js");
     const result = await deploy({
       appId: APP_ID,
       serverUrl: SERVER_URL,
@@ -391,7 +391,7 @@ describe("dev catalogue push behavior", () => {
     );
 
     const events: unknown[] = [];
-    const { pushMigration } = await import("./index.js");
+    const { pushMigration } = await import("./catalogue-project.js");
     const result = await pushMigration({
       appId: APP_ID,
       serverUrl: SERVER_URL,
@@ -417,8 +417,20 @@ describe("dev catalogue push behavior", () => {
     const migrationsDir = join(root, "migrations");
     await mkdir(migrationsDir, { recursive: true });
 
-    const fromHash = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
-    const toHash = "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb";
+    const fromSchema = {
+      users: s.table({
+        email: s.string(),
+      }),
+    };
+    const toSchema = {
+      users: s.table({
+        email_address: s.string(),
+      }),
+    };
+    const { computeSchemaHash } = await import("./catalogue.js");
+    const fromHash = await computeSchemaHash(s.defineApp(fromSchema).wasmSchema);
+    const toHash = await computeSchemaHash(s.defineApp(toSchema).wasmSchema);
+
     await writeFile(
       join(migrationsDir, `20260318-rename-${fromHash.slice(0, 12)}-${toHash.slice(0, 12)}.ts`),
       `
@@ -468,7 +480,7 @@ export default s.defineMigration({
       }),
     );
 
-    const { pushMigration } = await import("./index.js");
+    const { pushMigration } = await import("./catalogue-project.js");
     const result = await pushMigration({
       appId: APP_ID,
       serverUrl: SERVER_URL,
@@ -518,7 +530,7 @@ export default s.defineMigration({
     );
 
     const events: unknown[] = [];
-    const { pushSchema } = await import("./index.js");
+    const { pushSchema } = await import("./catalogue-project.js");
     const result = await pushSchema({
       appId: APP_ID,
       serverUrl: SERVER_URL,
@@ -576,7 +588,7 @@ export default s.defineMigration({
       }),
     );
 
-    const { pushPermissions } = await import("./index.js");
+    const { pushPermissions } = await import("./catalogue-project.js");
     const result = await pushPermissions({
       appId: APP_ID,
       serverUrl: SERVER_URL,
@@ -625,7 +637,7 @@ export default s.defineMigration({
     );
 
     const events: unknown[] = [];
-    const { deploy } = await import("./index.js");
+    const { deploy } = await import("./catalogue-project.js");
     const result = await deploy({
       appId: APP_ID,
       serverUrl: SERVER_URL,
@@ -713,7 +725,7 @@ export default s.defineMigration({
       }),
     );
 
-    const { deploy } = await import("./index.js");
+    const { deploy } = await import("./catalogue-project.js");
     const result = await deploy({
       appId: APP_ID,
       serverUrl: SERVER_URL,
