+++ packages/jazz-tools/src/dev/sveltekit.test.ts
@@ -3,6 +3,7 @@ import { mkdir, rm, writeFile } from "node:fs/promises";
 import { join } from "node:path";
 import { createTempRootTracker, getAvailablePort, todoSchema } from "./test-helpers.js";
 import * as devServer from "./dev-server.js";
+import * as catalogueProject from "./catalogue-project.js";
 import * as schemaWatcher from "./schema-watcher.js";
 import { jazzSvelteKit, __resetJazzSvelteKitPluginForTests } from "./sveltekit.js";
 import type { ViteDevServer } from "./vite.js";
@@ -140,7 +141,7 @@ describe("jazzSvelteKit", () => {
       backendSecret: "test-backend-secret",
       stop: vi.fn().mockResolvedValue(undefined),
     });
-    vi.spyOn(devServer, "deploy").mockResolvedValue(deployed("abc"));
+    vi.spyOn(catalogueProject, "deploy").mockResolvedValue(deployed("abc"));
     vi.spyOn(schemaWatcher, "watchSchema").mockReturnValue({ close: vi.fn() });
 
     const root = await tempRoots.create("jazz-sveltekit-top-level-telemetry-test-");
@@ -174,7 +175,7 @@ describe("jazzSvelteKit", () => {
       backendSecret: "test-backend-secret",
       stop: vi.fn().mockResolvedValue(undefined),
     });
-    vi.spyOn(devServer, "deploy").mockResolvedValue(deployed("abc"));
+    vi.spyOn(catalogueProject, "deploy").mockResolvedValue(deployed("abc"));
     vi.spyOn(schemaWatcher, "watchSchema").mockReturnValue({ close: vi.fn() });
 
     const root = await tempRoots.create("jazz-sveltekit-config-hook-");
@@ -209,7 +210,7 @@ describe("jazzSvelteKit", () => {
       backendSecret: "test-backend-secret",
       stop: vi.fn().mockResolvedValue(undefined),
     });
-    vi.spyOn(devServer, "deploy").mockResolvedValue(deployed("abc"));
+    vi.spyOn(catalogueProject, "deploy").mockResolvedValue(deployed("abc"));
     vi.spyOn(schemaWatcher, "watchSchema").mockReturnValue({ close: vi.fn() });
 
     const root = await tempRoots.create("jazz-sveltekit-no-restart-");
@@ -261,7 +262,7 @@ describe("jazzSvelteKit", () => {
       backendSecret: "test-backend-secret",
       stop: vi.fn().mockResolvedValue(undefined),
     });
-    vi.spyOn(devServer, "deploy").mockResolvedValue(deployed("abc"));
+    vi.spyOn(catalogueProject, "deploy").mockResolvedValue(deployed("abc"));
     vi.spyOn(schemaWatcher, "watchSchema").mockReturnValue({ close: vi.fn() });
 
     const plugin = jazzSvelteKit({
@@ -282,7 +283,7 @@ describe("jazzSvelteKit", () => {
       backendSecret: "test-backend-secret",
       stop: vi.fn().mockResolvedValue(undefined),
     });
-    vi.spyOn(devServer, "deploy").mockResolvedValue(deployed("abc"));
+    vi.spyOn(catalogueProject, "deploy").mockResolvedValue(deployed("abc"));
     vi.spyOn(schemaWatcher, "watchSchema").mockReturnValue({ close: vi.fn() });
 
     const root = await tempRoots.create("jazz-sveltekit-jwks-test-");
@@ -317,7 +318,7 @@ describe("jazzSvelteKit", () => {
       backendSecret: "test-backend-secret",
       stop: vi.fn().mockResolvedValue(undefined),
     });
-    vi.spyOn(devServer, "deploy").mockResolvedValue(deployed("abc"));
+    vi.spyOn(catalogueProject, "deploy").mockResolvedValue(deployed("abc"));
     vi.spyOn(schemaWatcher, "watchSchema").mockReturnValue({ close: vi.fn() });
 
     try {
@@ -352,7 +353,7 @@ describe("jazzSvelteKit", () => {
     process.env.PUBLIC_JAZZ_APP_ID = "00000000-0000-0000-0000-000000000010";
 
     vi.spyOn(devServer, "startLocalJazzServer");
-    vi.spyOn(devServer, "deploy").mockResolvedValue(deployed("abc"));
+    vi.spyOn(catalogueProject, "deploy").mockResolvedValue(deployed("abc"));
     vi.spyOn(schemaWatcher, "watchSchema").mockReturnValue({ close: vi.fn() });
 
     const plugin = jazzSvelteKit({ adminSecret: "env-test-admin" });
@@ -360,7 +361,7 @@ describe("jazzSvelteKit", () => {
     await (plugin.configureServer as (s: ViteDevServer) => Promise<void>)(viteServer);
 
     expect(devServer.startLocalJazzServer).not.toHaveBeenCalled();
-    expect(devServer.deploy).toHaveBeenCalledWith(
+    expect(catalogueProject.deploy).toHaveBeenCalledWith(
       expect.objectContaining({
         serverUrl: "http://jazz-test-server:4000",
         appId: "00000000-0000-0000-0000-000000000010",
@@ -371,7 +372,7 @@ describe("jazzSvelteKit", () => {
 
   it("connects to an existing server via options.server string URL", async () => {
     vi.spyOn(devServer, "startLocalJazzServer");
-    vi.spyOn(devServer, "deploy").mockResolvedValue(deployed("abc"));
+    vi.spyOn(catalogueProject, "deploy").mockResolvedValue(deployed("abc"));
     vi.spyOn(schemaWatcher, "watchSchema").mockReturnValue({ close: vi.fn() });
 
     const plugin = jazzSvelteKit({
@@ -382,7 +383,7 @@ describe("jazzSvelteKit", () => {
     await (plugin.configureServer as (s: ViteDevServer) => Promise<void>)(makeViteServer("serve"));
 
     expect(devServer.startLocalJazzServer).not.toHaveBeenCalled();
-    expect(devServer.deploy).toHaveBeenCalledWith(
+    expect(catalogueProject.deploy).toHaveBeenCalledWith(
       expect.objectContaining({
         serverUrl: "http://explicit-server:5000",
         appId: "00000000-0000-0000-0000-000000000020",
@@ -429,7 +430,7 @@ describe("jazzSvelteKit", () => {
       backendSecret: "test-backend-secret",
       stop: vi.fn().mockResolvedValue(undefined),
     });
-    vi.spyOn(devServer, "deploy").mockResolvedValue(deployed("abc123def4567890"));
+    vi.spyOn(catalogueProject, "deploy").mockResolvedValue(deployed("abc123def4567890"));
     let capturedOnPush: ((hash: string) => void) | undefined;
     vi.spyOn(schemaWatcher, "watchSchema").mockImplementation((opts) => {
       capturedOnPush = opts.onPush;
@@ -466,7 +467,7 @@ describe("jazzSvelteKit", () => {
       backendSecret: "test-backend-secret",
       stop: vi.fn().mockResolvedValue(undefined),
     });
-    vi.spyOn(devServer, "deploy").mockRejectedValue(new Error("schema push failed"));
+    vi.spyOn(catalogueProject, "deploy").mockRejectedValue(new Error("schema push failed"));
     vi.spyOn(schemaWatcher, "watchSchema").mockReturnValue({ close: vi.fn() });
 
     const root = await tempRoots.create("jazz-sveltekit-hmr-test-");
