+++ packages/jazz-tools/src/runtime/napi.integration.test.ts
@@ -10,6 +10,7 @@ import { type BatchId, type Row } from "./client.js";
 import type { Db, QueryBuilder, TableProxy } from "./db.js";
 import { translateQuery } from "./query-adapter.js";
 import { loadCompiledSchema, type LoadedSchemaProject } from "../schema-loader.js";
+import { deploy as deployProject } from "../dev/catalogue-project.js";
 import { deploy, startLocalJazzServer, startTestJwtIssuer } from "../testing/index.js";
 import { encodeSchema as encodeNativeSchema } from "./native-runtime/schema-codec.js";
 import {
@@ -267,7 +268,6 @@ function makePolicyTodoByIdQuery(schema: WasmSchema, id: string): QueryBuilder<P
   };
 }
 
-const BASIC_SCHEMA_DIR = fileURLToPath(new URL("../testing/fixtures/basic", import.meta.url));
 const TODO_SERVER_SCHEMA_DIR = fileURLToPath(
   new URL("../../../../examples/todo-server-ts", import.meta.url),
 );
@@ -539,13 +539,14 @@ describe("NAPI integration", () => {
     try {
       const { createJazzContext } = await import("../backend/create-jazz-context.js");
 
+      const todoServerProject = await loadTodoServerProject();
       await deploy({
         serverUrl: server.url,
         appId,
         adminSecret,
-        schemaDir: TODO_SERVER_SCHEMA_DIR,
+        schema: todoServerProject.wasmSchema,
+        permissions: todoServerProject.permissions,
       });
-      const todoServerProject = await loadTodoServerProject();
       const todoServerSchema = todoServerProject.wasmSchema;
       const policyTodosTable = makePolicyTodosTable(todoServerSchema);
 
@@ -684,7 +685,7 @@ describe("NAPI integration", () => {
     try {
       const { createJazzContext } = await import("../backend/create-jazz-context.js");
 
-      await deploy({
+      await deployProject({
         serverUrl: server.url,
         appId,
         adminSecret,
@@ -824,7 +825,7 @@ describe("NAPI integration", () => {
     } | null = null;
 
     try {
-      await deploy({
+      await deployProject({
         serverUrl: server.url,
         appId,
         adminSecret,
@@ -1077,7 +1078,7 @@ describe("NAPI integration", () => {
         adminSecret,
         dataDir,
       });
-      await deploy({
+      await deployProject({
         serverUrl: server.url,
         appId,
         adminSecret,
@@ -1132,7 +1133,7 @@ describe("NAPI integration", () => {
         dataDir,
         jwksUrl: jwtIssuer.jwksUrl,
       });
-      await deploy({
+      await deployProject({
         serverUrl: server.url,
         appId,
         adminSecret,
@@ -1386,7 +1387,7 @@ describe("NAPI integration", () => {
         `,
       );
 
-      await deploy({
+      await deployProject({
         serverUrl: server.url,
         appId,
         adminSecret,
@@ -1417,13 +1418,14 @@ describe("NAPI integration", () => {
     try {
       const { createJazzContext } = await import("../backend/create-jazz-context.js");
 
+      const todoServerProject = await loadTodoServerProject();
       await deploy({
         serverUrl: server.url,
         appId,
         adminSecret,
-        schemaDir: TODO_SERVER_SCHEMA_DIR,
+        schema: todoServerProject.wasmSchema,
+        permissions: todoServerProject.permissions,
       });
-      const todoServerProject = await loadTodoServerProject();
       const todoServerSchema = todoServerProject.wasmSchema;
       const policyTodosTable = makePolicyTodosTable(todoServerSchema);
 
@@ -1505,7 +1507,7 @@ describe("NAPI integration", () => {
         serverUrl: server.url,
         appId,
         adminSecret,
-        schemaDir: BASIC_SCHEMA_DIR,
+        schema: TEST_SCHEMA,
       });
 
       writerRuntimeData = await createTempRuntimeData("jazz-napi-sync-writer-");
