+++ packages/jazz-tools/src/testing/index.test.ts
@@ -253,23 +253,9 @@ describe("startLocalJazzServer", () => {
 });
 
 describe("deploy", () => {
-  it("rejects when no root schema.ts can be found", async () => {
-    const root = await createTempRoot("jazz-tools-testing-missing-schema-");
-
-    await expect(
-      deploy({
-        serverUrl: "http://127.0.0.1:9999",
-        appId: "00000000-0000-0000-0000-000000000001",
-        adminSecret: "admin-secret",
-        schemaDir: root,
-      }),
-    ).rejects.toThrow(/schema file not found/i);
-  });
-
-  it("deploys the current schema object via schema.ts", async () => {
+  it("deploys the current schema object", async () => {
     const port = await getAvailablePort();
     const adminSecret = "admin-secret";
-    const schemaDir = join(import.meta.dirname, "fixtures/basic");
 
     const server = await startTrackedLocalJazzServer({
       appId: "00000000-0000-0000-0000-000000000001",
@@ -282,7 +268,8 @@ describe("deploy", () => {
         serverUrl: server.url,
         appId: "00000000-0000-0000-0000-000000000001",
         adminSecret,
-        schemaDir,
+        schema: testApp,
+        permissions: testPermissions,
       });
 
       expect(result.schema.hash).toBeTruthy();
@@ -302,14 +289,13 @@ describe("deploy", () => {
   }, 30_000);
 
   it("rejects when server is unreachable", async () => {
-    const schemaDir = join(import.meta.dirname, "fixtures/basic");
-
     await expect(
       deploy({
         serverUrl: "http://127.0.0.1:9",
         appId: "00000000-0000-0000-0000-000000000001",
         adminSecret: "admin-secret",
-        schemaDir,
+        schema: testApp,
+        permissions: testPermissions,
       }),
     ).rejects.toThrow();
   }, 10_000);
