+++ packages/jazz-tools/tests/browser/db.disconnect.test.ts
@@ -1,13 +1,9 @@
 import { afterEach, describe, expect, it } from "vitest";
-import { CompiledPermissions, schema as s } from "../../src/";
+import { schema as s } from "../../src/";
 import { createDb, Db, type QueryBuilder } from "../../src/runtime/db.js";
 import { generateAuthSecret } from "../../src/runtime/auth-secret-store.js";
 import { PersistentBrowserOpfsRuntime } from "../../src/runtime/native-runtime/persistent-browser-runtime.js";
-import {
-  fetchPermissionsHead,
-  publishStoredPermissions,
-  publishStoredSchema,
-} from "../../src/runtime/schema-fetch.js";
+import { deploy } from "../../src/dev/catalogue.js";
 import { TestCleanup, sleep, uniqueDbName, waitForQuery, withTimeout } from "./support.js";
 import { getJazzServerInfo, type JazzServerInfo } from "./testing-server.js";
 
@@ -469,31 +465,17 @@ async function waitForTodos(
   return waitForQuery(db, app.todos, predicate, label, timeoutMs, tier);
 }
 
-async function publishSyncServerSchemaAndPermissions(appId: string): Promise<JazzServerInfo> {
-  const testingServer = await getJazzServerInfo(appId);
-  await publishPermissionsForServer(testingServer, allowAllPermissions);
-  return testingServer;
-}
-
-async function publishPermissionsForServer(
-  testingServer: JazzServerInfo,
-  permissions: CompiledPermissions,
-): Promise<void> {
+async function publishSyncServerSchemaAndPermissions(
+  requestedAppId: string,
+): Promise<JazzServerInfo> {
+  const testingServer = await getJazzServerInfo(requestedAppId);
   const { appId, serverUrl, adminSecret } = testingServer;
-  const { hash: schemaHash } = await publishStoredSchema(serverUrl, {
+  await deploy({
     appId,
+    serverUrl,
     adminSecret,
-    schema: app.wasmSchema,
+    schema: app,
+    permissions: allowAllPermissions,
   });
-  const { head } = await fetchPermissionsHead(serverUrl, {
-    appId,
-    adminSecret,
-  });
-  await publishStoredPermissions(serverUrl, {
-    appId,
-    adminSecret,
-    schemaHash,
-    permissions,
-    expectedParentBundleObjectId: head?.bundleObjectId ?? null,
-  });
+  return testingServer;
 }
